+++ src/FormElement/Number.php
@@ -2,6 +2,8 @@
 
 namespace Xtreamwayz\HTMLFormValidator\FormElement;
 
+use InvalidArgumentException;
+
 class Number extends AbstractFormElement
 {
     /**
@@ -16,37 +18,43 @@ class Number extends AbstractFormElement
      */
     protected function attachDefaultValidators()
     {
-        $baseValue = 0;
-        $min = $this->element->getAttribute('min');
-        $max = $this->element->getAttribute('max');
-        $step = $this->element->getAttribute('step');
+        $step = ($this->element->hasAttribute('step')) ? $this->element->getAttribute('step') : 1;
 
-        $this->attachValidatorByName('isint');
+        var_dump('step:' . $step .':'. (int) $step);
 
-        if ($min && $max) {
-            $baseValue = $min;
-            $this->attachValidatorByName('between', [
-                'min'       => $min,
-                'max'       => $max,
-                'inclusive' => true,
+        if (is_numeric($step) && (int) $step == $step) {
+            var_dump(__LINE__);
+            $this->attachValidatorByName('isint');
+        } elseif (is_numeric($step) && (float) $step == $step) {
+            var_dump(__LINE__);
+            $this->attachValidatorByName('isfloat', [
+                'locale' => 'en'
             ]);
-        } elseif ($min) {
-            $baseValue = $min;
+        } elseif ($step != 'any') {
+            var_dump(__LINE__);
+            throw new InvalidArgumentException('Number step must be an int, float or the text "any"');
+        }
+
+        if ($this->element->hasAttribute('min')) {
+            var_dump('min:' . $this->element->getAttribute('min'));
             $this->attachValidatorByName('greaterthan', [
-                'min'       => $min,
-                'inclusive' => true,
-            ]);
-        } elseif ($max) {
-            $this->attachValidatorByName('lessthan', [
-                'max'       => $max,
+                'min'       => $this->element->getAttribute('min'),
                 'inclusive' => true,
             ]);
         }
 
-        if ($step) {
+        if ($this->element->hasAttribute('max')) {
+            var_dump('max:' . $this->element->getAttribute('max'));
+            $this->attachValidatorByName('lessthan', [
+                'max'       => $this->element->getAttribute('max'),
+                'inclusive' => true,
+            ]);
+        }
+
+        if (is_numeric($step)) {
             $this->attachValidatorByName('step', [
                 'step'      => $step,
-                'baseValue' => $baseValue,
+                'baseValue' => ($this->element->hasAttribute('min')) ? $this->element->getAttribute('min') : 0,
             ]);
         }
     }